backend/requirements.txt
asttokens==2.4.1
async-timeout==4.0.3
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kaggle==1.5.16
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
numpy==1.26.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pika==1.3.2
Pillow==9.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.0
pymongo==4.6.0
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
text-unidecode==1.3
tokenizers==0.15.0
#torch==2.1.0+cpu
#torchaudio==2.1.0+cpu
#torchvision==0.16.0+cpu
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.10
webencodings==0.5.1
Werkzeug==3.0.1
